s3cfb_ops/init: fix LPM when using a JB bootloader
after commit 26f3de0 the device still rebooted when trying to charge it powered off. this change now fixes
this and LPM is finally working again. Problem on GT-N51x0 devices was that the KK bootloader uses a different
command line option for entering low power mode. So we gonna exchange the old with the new option before passing the
commandline to the init process

Credits to FennyFatal for the inital version for another device
and again Kumajaya for the s3cfb_ops changes